[FEATURE] improved __getitem__ code

+++ tests/test_files/minimal.pxd
@@ -49,7 +49,7 @@ cdef extern from "minimal.hpp":
         int operator()(Minimal) # wrap-cast:toInt
 
         int size()
-        int operator[](int) #wrap-upper-limit:size()
+        int operator[](size_t) #wrap-upper-limit:size()
 
         int sumup(libcpp_vector[int] what)
         int call(libcpp_vector[Minimal] what) # ref-arg-out:0
